+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add folder tree download
 - Add config validation
 - Various documentation and logging improvements
+- Remove remaining hardcoded linux-style paths
 
 # 0.1.1-beta